+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Additions
 
   * `types::UserState::Deactivated` has been added.
+  * New endpoints for:
+    - Sharing and unsharing groups
+    - Listing shared projects in a group
 
 # v0.1500.0